python-api/requirements.txt
agno>=1.4.4
anthropic>=0.46.0
fastapi>=0.135.0
httpx>=0.28.1
mcp>=1.0.0
openpyxl>=3.1.5
pandas>=2.2.3
pyjwt>=2.10.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
pydantic>=2.10.6
pydantic-settings>=2.8.1
pytest>=8.3.5
supabase>=2.15.0
uvicorn>=0.34.0
